+++ common/src/main/java/dev/tocraft/ctgen/data/SurfaceBuilderAccess.java
@@ -9,10 +9,11 @@ import net.minecraft.world.level.levelgen.NoiseChunk;
 import net.minecraft.world.level.levelgen.RandomState;
 import net.minecraft.world.level.levelgen.SurfaceRules;
 import net.minecraft.world.level.levelgen.WorldGenerationContext;
+import net.minecraft.world.level.levelgen.synth.SimplexNoise;
 
 import java.util.function.Supplier;
 
 public interface SurfaceBuilderAccess {
-    void ctgen$buildSurface(RandomState noiseConfig, BiomeManager biomeAccess, Registry<Biome> biomeRegistry, boolean useLegacyRandom, WorldGenerationContext heightContext, final ChunkAccess chunk, NoiseChunk chunkNoiseSampler, SurfaceRules.RuleSource materialRule, Supplier<MapSettings> mapSettings);
+    void ctgen$buildSurface(RandomState noiseConfig, BiomeManager biomeAccess, Registry<Biome> biomeRegistry, boolean useLegacyRandom, WorldGenerationContext heightContext, final ChunkAccess chunk, NoiseChunk chunkNoiseSampler, SurfaceRules.RuleSource materialRule, Supplier<MapSettings> mapSettings, Supplier<SimplexNoise> noise);
 
 }