+++ common/src/main/java/dev/tocraft/ctgen/mixin/SurfaceSystemMixin.java
@@ -13,6 +13,7 @@ import net.minecraft.world.level.levelgen.RandomState;
 import net.minecraft.world.level.levelgen.SurfaceRules;
 import net.minecraft.world.level.levelgen.SurfaceSystem;
 import net.minecraft.world.level.levelgen.WorldGenerationContext;
+import net.minecraft.world.level.levelgen.synth.SimplexNoise;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
@@ -28,16 +29,19 @@ public abstract class SurfaceSystemMixin implements SurfaceBuilderAccess {
 
     @Unique
     private Supplier<MapSettings> ctgen$settings = () -> null;
+    private Supplier<SimplexNoise> ctgen$noise = () -> null;
 
     @Unique
     @Override
-    public void ctgen$buildSurface(RandomState randomState, BiomeManager biomeManager, Registry<Biome> biomes, boolean useLegacyRandomSource, WorldGenerationContext context, final ChunkAccess chunk, NoiseChunk noiseChunk, SurfaceRules.RuleSource ruleSource, Supplier<MapSettings> mapSettings) {
+    public void ctgen$buildSurface(RandomState randomState, BiomeManager biomeManager, Registry<Biome> biomes, boolean useLegacyRandomSource, WorldGenerationContext context, final ChunkAccess chunk, NoiseChunk noiseChunk, SurfaceRules.RuleSource ruleSource, Supplier<MapSettings> mapSettings, Supplier<SimplexNoise> noise) {
         this.ctgen$settings = mapSettings;
+        this.ctgen$noise = noise;
         this.buildSurface(randomState, biomeManager, biomes, useLegacyRandomSource, context, chunk, noiseChunk, ruleSource);
     }
 
     @Inject(method = "buildSurface", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/levelgen/SurfaceRules$RuleSource;apply(Ljava/lang/Object;)Ljava/lang/Object;"))
     private void setMapContext(RandomState randomState, BiomeManager biomeManager, Registry<Biome> biomes, boolean useLegacyRandomSource, WorldGenerationContext context, ChunkAccess chunk, NoiseChunk noiseChunk, SurfaceRules.RuleSource ruleSource, CallbackInfo ci, @Local SurfaceRules.Context context2) {
         ((MapInfoAccessor) (Object) context2).ctgen$setSettings(this.ctgen$settings);
+        ((MapInfoAccessor) (Object) context2).ctgen$setNoise(this.ctgen$noise);
     }
 }